requested changes
- Add plain-text fallback for non-kind-14 messages to prevent blank rendering
- Unify content extraction between bubble and snippet views using shared utility
- Add URL scheme validation (http/https only) for nested images
- Make parse error warnings development-only to reduce log noise
- Extract reusable functions to src/lib/utils/message-content.ts

Fixes:
- Restores non-kind-14 plain-text fallback when no structured renderer matches
- Ensures snippet/bubble parity for kind-14 messages with JSON-wrapped content
- Prevents rendering of malicious image URLs (javascript:, data:, etc.)
- Eliminates parse warnings on adversarial relay data in production"
